@@ -1,6 +1,6 @@
 {
   "name": "samlify",
-  "version": "2.7.6",
+  "version": "2.8.2",
   "description": "High-level API for Single Sign On (SAML 2.0)",
   "main": "build/index.js",
   "keywords": [
@@ -31,31 +31,32 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@authenio/xml-encryption": "^1.2.2",
-    "camelcase": "^5.3.1",
+    "@authenio/xml-encryption": "^1.3.0",
+    "@xmldom/xmldom": "^0.7.4",
+    "camelcase": "^6.2.0",
     "node-forge": "^0.10.0",
-    "node-rsa": "^1.0.5",
+    "node-rsa": "^1.1.1",
     "pako": "^1.0.10",
     "uuid": "^3.3.2",
     "xml": "^1.0.1",
-    "xml-crypto": "^2.1.0",
-    "xmldom": "^0.5.0",
-    "xpath": "^0.0.27"
+    "xml-crypto": "^2.1.3",
+    "xpath": "^0.0.32"
   },
   "devDependencies": {
     "@ava/typescript": "^1.1.1",
     "@types/node": "^11.11.3",
-    "@types/node-forge": "^0.9.5",
+    "@types/node-forge": "^0.10.3",
+    "@types/node-rsa": "^1.1.1",
     "@types/pako": "^1.0.1",
     "@types/uuid": "3.0.0",
-    "@types/xmldom": "^0.1.28",
-    "ava": "^3.8.2",
+    "@types/xmldom": "^0.1.31",
+    "ava": "^3.15.0",
     "coveralls": "^3.1.0",
     "nyc": "^15.0.1",
     "timekeeper": "^2.2.0",
     "ts-node": "^8.3.0",
     "tslint": "^6.1.2",
-    "typescript": "^3.8.3"
+    "typescript": "^4.4.2"
   },
   "ava": {
     "extensions": [
